refactor: 에러 메시지 UI 를 별도의 컴포넌트로 분리
