+++ src/pages/HomePage/components/Category/Category.styles.ts
@@ -31,18 +31,3 @@ export const ThemeGrid = styled.div`
   margin-top: ${({ theme }) => theme.spacing[5]};
   margin-bottom: ${({ theme }) => theme.spacing[5]};
 `;
-
-export const ErrorContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  margin-top: 124px;
-  margin-bottom: 124px;
-`;
-
-export const ErrorMessage = styled.p`
-  font-size: ${({ theme }) => theme.typography.label.label1Regular.fontSize};
-  font-weight: ${({ theme }) =>
-    theme.typography.label.label1Regular.fontWeight};
-`;
