+++ src/pages/HomePage/components/TrendingGifts/TrendingGifts.styles.tsx
@@ -90,19 +90,3 @@ export const LoadingContainer = styled.div`
   margin-top: 124px;
   margin-bottom: 124px;
 `;
-
-export const ErrorContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 124px;
-  margin-bottom: 124px;
-`;
-
-export const ErrorMessage = styled.p`
-  font-size: ${({ theme }) => theme.typography.label.label1Regular.fontSize};
-  font-weight: ${({ theme }) =>
-    theme.typography.label.label1Regular.fontWeight};
-`;
